backend/requirements.txt
flask
torch
torchvision
opencv-python-headless
mtcnn
pillow
werkzeug
dlib
efficientnet_pytorch
pretrainedmodels
numpy
flask-cors
requests
gunicorn